move network panel to its own file
